Fix the vector access function to use size_t
As in the C++ standard library, vectors use
size_t as the type of the variable used to
index the array. This makes sense since when
using a for loop, you need to check the index
against vector_size, which is of type size_t.
Hence, the index needs to be of type size_t.
Hence the vector access variable needs to be
of size_t as well.

vector.h:
 * modify vector access function header to use size_t
   vs unsigned int.

vector.c:
 * modify vector access function to use size_t
   vs unsigned int.

+++ vector.c
@@ -24,7 +24,7 @@ vector_ctor
 // Can set the errno when out of range.
 void *
 vector_at
-(struct vector *v, unsigned int pos) {
+(struct vector *v, size_t pos) {
 	return v->arr[pos];
 }
