+++ vector.h
@@ -11,7 +11,7 @@ vector_ctor
 // Can set the errno when out of range.
 void *
 vector_at
-(struct vector *v, unsigned int pos);
+(struct vector *v, size_t pos);
 
 // We will want to implement something like;
 // operator[] at some point.
